ver_gpt-4o/requirements.txt
functions-framework==3.*
requests
openai
python-dotenv
firebase-admin
configparser
tiktoken
langchain
faiss-cpu